x11/qt5-qtwebengine: Use mk/atomic64.mk.
On i386, this fails with, e.g.:

/usr/bin/ld: obj/v8/v8_base_without_compiler/api.o: in function `v8::Isolate::AdjustAmountOfExternalAllocatedMemory(long long)':
api.cc:(.text._ZN2v87Isolate37AdjustAmountOfExternalAllocatedMemoryEx+0x28): undefined reference to `__atomic_fetch_add_8'
/usr/bin/ld: api.cc:(.text._ZN2v87Isolate37AdjustAmountOfExternalAllocatedMemoryEx+0x48): undefined reference to `__atomic_load_8'

It's probably going to fail to build in the same way on all platforms
without native 64-bit atomics.  On platforms with native 64-bit
atomics, mk/atomic64.mk has no effect, so this should be harmless --
let's just flip it on and see if that helps with 32-bit platforms.